Refactor: extract type definitions and type helpers from compiler.c
Move all internal compiler type definitions (250 lines of typedefs and
structs) into src/internal/basl_compiler_types.h, and extract 31 pure
type-helper functions into src/compiler_types.c (389 lines).

Extracted functions:
- 21 type predicate functions (basl_parser_type_is_*)
- Type comparison and assignability checks
- Unary/binary operator type support checks
- Function/interface return type accessors
- basl_program_function_type_decl
- basl_class_decl_implements_interface

compiler.c: 20,359 → 19,714 lines (-645)

This is a pure mechanical extraction — no logic changes. The shared
header enables future extraction of additional compiler subsystems
(string parsing, built-in dispatch, type reference parsing) that
depend on the same type definitions.
